+++ src/app/(routes)/login/page.tsx
@@ -1,14 +1,9 @@
 "use client";
 
-import { getGoogleAuthURL } from "@/api/oauth/google";
-import GoogleLogo from "@/assets/icons/google.svg";
+import GoogleLoginButton from "@/components/atoms/Button/GoogleLoginButton";
 import KakaoLoginButton from "@/components/atoms/Button/KakaoLoginButton";
 
 export default function LoginPage() {
-  const handleGoogleLogin = () => {
-    window.location.href = getGoogleAuthURL();
-  };
-
   return (
     <div className="flex h-screen flex-col items-center justify-center gap-16">
       <div className="flex flex-col items-center justify-center gap-[1.375rem]">
@@ -20,15 +15,7 @@ export default function LoginPage() {
       <div className="flex flex-col items-center justify-center gap-4">
         <KakaoLoginButton />
 
-        <button
-          onClick={handleGoogleLogin}
-          className="relative flex h-12 w-[19.8125rem] cursor-pointer items-center justify-center rounded-xl border-[0.8px] border-[#B0B0B0] bg-white transition-all"
-        >
-          <GoogleLogo className="mr-3 h-5 w-5" />
-          <p className="font-['Roboto'] text-sm font-medium text-[#1f1f1f]">
-            Google 계정으로 로그인
-          </p>
-        </button>
+        <GoogleLoginButton />
       </div>
     </div>
   );
